+++ src/Server/App/WampApplication.php
@@ -77,9 +77,13 @@ class WampApplication implements PushableWampServerInterface, LoggerAwareInterfa
     /**
      * @param string|array $data
      * @param string       $provider
+     *
+     * @deprecated to be removed in 4.0, use the symfony/messenger component instead
      */
     public function onPush(WampRequest $request, $data, $provider): void
     {
+        trigger_deprecation('gos/web-socket-bundle', '3.7', '%s() is deprecated and will be removed in 4.0, use the symfony/messenger component instead.', __METHOD__);
+
         if (null !== $this->logger) {
             $this->logger->info(
                 sprintf('Pusher %s has pushed', $provider),